+++ packages/mtbird-editor/src/components/Toolbar/ToolBox/index.tsx
@@ -1,16 +1,20 @@
 import React, { useContext } from 'react';
-import type { IComponent } from '@mtbird/shared';
+import type { IComponent, IComponentInstance } from '@mtbird/shared';
 import styles from './style.module.less';
 import models from '../../../store/types';
 
-export default ({ component }: IComponent) => {
+interface IProps {
+  component: IComponent<IComponentInstance>;
+}
+
+export default ({ component }: IProps) => {
   const { addComponent } = useContext(models).actions;
   const handleClick = () => {
     addComponent(component.instance);
   };
 
   return (
-    <div className={styles.toolbarBoxContainer} onClick={handleClick}>
+    <div className={styles.toolbarBoxContainer} onClick={handleClick} id={`${component.componentName}Box`}>
       {component.icon.startsWith('http') ? (
         <img className={styles.toolbarBoxIcon} src={component.icon} />
       ) : (